@@ -1,6 +1,6 @@
 {
   "name": "serverless-domain-manager",
-  "version": "6.4.4",
+  "version": "7.0.0",
   "engines": {
     "node": ">=14"
   },
@@ -27,10 +27,10 @@
   "main": "dist/src/index.js",
   "bin": {},
   "scripts": {
-    "integration-basic": "nyc mocha -r ts-node/register --project tsconfig.json --parallel test/integration-tests/basic.test.ts",
-    "integration-deploy": "nyc mocha -r ts-node/register --project tsconfig.json --parallel test/integration-tests/deploy.test.ts",
-    "integration-debug": "nyc mocha -r ts-node/register --project tsconfig.json test/integration-tests/debug.test.ts",
-    "test": "nyc mocha -r ts-node/register --project tsconfig.json test/unit-tests/index.test.ts && nyc report --reporter=text-summary",
+    "integration-basic": "nyc mocha -r ts-node/register --project tsconfig.json --parallel test/integration-tests/basic/basic.test.ts",
+    "integration-deploy": "nyc mocha -r ts-node/register --project tsconfig.json --parallel test/integration-tests/deploy/deploy.test.ts",
+    "integration-debug": "nyc mocha -r ts-node/register --project tsconfig.json test/integration-tests/debug/debug.test.ts",
+    "test": "find ./test/unit-tests -name '*.test.ts' | xargs nyc mocha -r ts-node/register --project tsconfig.json --timeout 5000 && nyc report --reporter=text-summary",
     "test:debug": "NODE_OPTIONS='--inspect-brk' mocha -j 1 -r ts-node/register --project tsconfig.json test/unit-tests/index.test.ts",
     "integration-test": "npm run integration-basic && npm run integration-deploy",
     "lint": "tslint --project . && tslint --project tsconfig.json",
@@ -49,22 +49,35 @@
   },
   "devDependencies": {
     "@types/mocha": "^10.0.1",
-    "@types/node": "^18.14.1",
-    "aws-sdk-mock": "^5.8.0",
+    "@types/node": "^18.15.5",
+    "@types/shelljs": "^0.8.11",
+    "aws-sdk-client-mock": "^2.1.1",
     "chai": "^4.3.7",
     "chai-spies": "^1.0.0",
     "mocha": "^10.2.0",
     "mocha-param": "^2.0.1",
     "nyc": "^15.1.0",
     "randomstring": "^1.2.3",
-    "serverless": "^3.27.0",
+    "serverless": "^3.28.1",
     "serverless-plugin-split-stacks": "^1.12.0",
     "shelljs": "^0.8.5",
     "ts-node": "^10.9.1",
     "tslint": "^6.1.3",
-    "typescript": "^4.9.5"
+    "typescript": "^5.0.2"
   },
   "dependencies": {
-    "aws-sdk": "^2.1322.0"
+    "@aws-sdk/client-acm": "^3.295.0",
+    "@aws-sdk/client-api-gateway": "^3.295.0",
+    "@aws-sdk/client-apigatewayv2": "^3.295.0",
+    "@aws-sdk/client-cloudformation": "^3.295.0",
+    "@aws-sdk/client-route-53": "^3.295.0",
+    "@aws-sdk/client-s3": "^3.295.0",
+    "@aws-sdk/config-resolver": "^3.295.0",
+    "@aws-sdk/credential-providers": "^3.295.0",
+    "@aws-sdk/node-config-provider": "^3.295.0",
+    "@aws-sdk/smithy-client": "^3.295.0"
+  },
+  "peerDependencies": {
+    "serverless": "^2.60 || 3"
   }
 }
