@@ -1,6 +1,6 @@
 import "mocha";
-import utilities = require("./test-utilities");
-import {TEMP_DIR} from "./base";
+import utilities = require("../test-utilities");
+import {TEMP_DIR} from "../base";
 
 const CONFIGS_FOLDER = "debug";
 const TIMEOUT_MINUTES = 15 * 60 * 1000; // 15 minutes in milliseconds
