@@ -1,13 +1,16 @@
-import chai = require("chai");
 import "mocha";
+import chai = require("chai");
 import itParam = require("mocha-param");
-
-import utilities = require("./test-utilities");
-import {TEST_DOMAIN, PLUGIN_IDENTIFIER, RANDOM_STRING} from "./base";// tslint:disable-line
+import shell = require("shelljs");
+import utilities = require("../test-utilities");
+import APIGatewayWrap from "../apigateway";
+import {TEST_DOMAIN, PLUGIN_IDENTIFIER, RANDOM_STRING} from "../base";
 
 const expect = chai.expect;
 const CONFIGS_FOLDER = "deploy";
 const TIMEOUT_MINUTES = 10 * 60 * 1000; // 10 minutes in milliseconds
+// the us-west-2 is set in each test config
+const apiGatewayClient = new APIGatewayWrap("us-west-2");
 
 const testCases = [
     {
@@ -27,7 +30,6 @@ const testCases = [
         testStage: "test",
     },
     {
-        createApiGateway: true,
         restApiName: "rest-api-custom",
         testBasePath: "(none)",
         testDescription: "Enabled with custom api gateway",
@@ -117,13 +119,13 @@ const testCases = [
         testStage: "test",
     },
     {
-      testBasePath: "(none)",
-      testDescription: "Deploy with split horizon dns",
-      testDomain: `${PLUGIN_IDENTIFIER}-split-horizon-dns-${RANDOM_STRING}.${TEST_DOMAIN}`,
-      testEndpoint: "REGIONAL",
-      testFolder: `${CONFIGS_FOLDER}/split-horizon-dns`,
-      testStage: "test",
-  },
+        testBasePath: "(none)",
+        testDescription: "Deploy with split horizon dns",
+        testDomain: `${PLUGIN_IDENTIFIER}-split-horizon-dns-${RANDOM_STRING}.${TEST_DOMAIN}`,
+        testEndpoint: "REGIONAL",
+        testFolder: `${CONFIGS_FOLDER}/split-horizon-dns`,
+        testStage: "test",
+    },
 ];
 
 describe("Integration Tests", function () {
@@ -133,23 +135,29 @@ describe("Integration Tests", function () {
         // @ts-ignore
         itParam("${value.testDescription}", testCases, async (value) => {
             let restApiInfo;
-            if (value.createApiGateway) {
-                restApiInfo = await utilities.setupApiGatewayResources(value.restApiName);
+            if (value.restApiName) {
+                restApiInfo = await apiGatewayClient.setupApiGatewayResources(value.restApiName);
+
+                shell.env.REST_API_ID = restApiInfo.restApiId;
+                shell.env.RESOURCE_ID = restApiInfo.resourceId;
             }
             try {
                 await utilities.createResources(value.testFolder, value.testDomain);
-                const stage = await utilities.getStage(value.testDomain);
+                const stage = await apiGatewayClient.getStage(value.testDomain);
                 expect(stage).to.equal(value.testStage);
 
-                const basePath = await utilities.getBasePath(value.testDomain);
+                const basePath = await apiGatewayClient.getBasePath(value.testDomain);
                 expect(basePath).to.equal(value.testBasePath);
 
-                const endpoint = await utilities.getEndpointType(value.testDomain);
+                const endpoint = await apiGatewayClient.getEndpointType(value.testDomain);
                 expect(endpoint).to.equal(value.testEndpoint);
             } finally {
                 await utilities.destroyResources(value.testDomain);
-                if (value.createApiGateway) {
-                    await utilities.deleteApiGatewayResources(restApiInfo.restApiId);
+                if (value.restApiName) {
+                    await apiGatewayClient.deleteApiGatewayResources(restApiInfo.restApiId);
+
+                    delete shell.env.REST_API_ID;
+                    delete shell.env.RESOURCE_ID;
                 }
             }
         });
