@@ -1,5 +1,5 @@
-import DomainConfig = require("./models/domain-config");
 import {ServerlessInstance, ServerlessOptions, ServerlessUtils} from "./types";
+import {fromIni} from "@aws-sdk/credential-providers";
 
 export default class Globals {
 
@@ -8,6 +8,8 @@ export default class Globals {
     public static serverless: ServerlessInstance;
     public static options: ServerlessOptions;
     public static v3Utils: ServerlessUtils;
+
+    public static currentRegion: string;
 
     public static defaultRegion = "us-east-1";
     public static defaultBasePath = "(none)";
@@ -49,6 +51,7 @@ export default class Globals {
     public static tlsVersions = {
         tls_1_0: "TLS_1_0",
         tls_1_2: "TLS_1_2",
+        tls_1_3: "TLS_1_3",
     };
 
     public static routingPolicies = {
@@ -61,66 +64,7 @@ export default class Globals {
         return Globals.options.stage || Globals.serverless.service.provider.stage;
     }
 
-    public static cliLog(prefix: string, message: string): void {
-        Globals.serverless.cli.log(`${prefix} ${message}`, Globals.pluginName);
-    }
-
-    /**
-     * Logs error message
-     */
-    public static logError(message: string): void {
-        if (Globals.v3Utils) {
-            Globals.v3Utils.log.error(message);
-        } else {
-            Globals.cliLog("[Error]", message);
-        }
-    }
-
-    /**
-     * Logs info message
-     */
-    public static logInfo(message: string): void {
-        if (Globals.v3Utils) {
-            Globals.v3Utils.log.verbose(message);
-        } else {
-            Globals.cliLog("[Info]", message);
-        }
-    }
-
-    /**
-     * Logs warning message
-     */
-    public static logWarning(message: string): void {
-        if (Globals.v3Utils) {
-            Globals.v3Utils.log.warning(message);
-        } else {
-            Globals.cliLog("[WARNING]", message);
-        }
-    }
-
-    /**
-     * Prints out a summary of all domain manager related info
-     */
-    public static printDomainSummary(domains: DomainConfig[]): void {
-        const summaryList = [];
-        domains.forEach((domain) => {
-            if (domain.domainInfo) {
-                summaryList.push(`Domain Name: ${domain.givenDomainName}`);
-                summaryList.push(`Target Domain: ${domain.domainInfo.domainName}`);
-                summaryList.push(`Hosted Zone Id: ${domain.domainInfo.hostedZoneId}`);
-            }
-        });
-        // don't print summary if summaryList is empty
-        if (!summaryList.length) {
-            return;
-        }
-        if (Globals.v3Utils) {
-            Globals.serverless.addServiceOutputSection(Globals.pluginName, summaryList);
-        } else {
-            Globals.cliLog("[Summary]", "Distribution Domain Name");
-            summaryList.forEach((item) => {
-                Globals.cliLog("", `${item}`);
-            });
-        }
+    public static async getProfileCreds(profile: string) {
+        return await fromIni({profile})();
     }
 }
