@@ -1,5 +1,3 @@
-import {HTTPOptions} from "aws-sdk";
-
 export interface Route53Params {
     routingPolicy: "simple" | "latency" | "weighted" | undefined;
     weight: number | undefined;
@@ -47,8 +45,8 @@ export interface ServerlessInstance {
                 Outputs: any,
             },
             apiGateway: {
-                restApiId: string,
-                websocketApiId: string,
+                restApiId: any,
+                websocketApiId: any,
             },
             tags: Tags,
             stackTags: Tags,
@@ -58,28 +56,8 @@ export interface ServerlessInstance {
             customDomains?: CustomDomain[],
         },
     };
-    providers: {
-        aws: {
-            sdk: {
-                APIGateway: any,
-                ApiGatewayV2: any,
-                Route53: any,
-                CloudFormation: any,
-                ACM: any,
-                S3: any,
-                config: {
-                    httpOptions: HTTPOptions,
-                    update(toUpdate: object): void,
-                },
-                SharedIniFileCredentials: any,
-            }
-            getCredentials(),
-            getRegion(),
-        },
-    };
     cli: {
-        log(str: string, entity?: string),
-        consoleLog(str: any),
+        log(str: string, entity?: string)
     };
 
     addServiceOutputSection?(name: string, data: string[]);
@@ -101,7 +79,7 @@ export interface ServerlessProgressFactory {
 
 export interface ServerlessUtils {
     writeText: (message: string) => void,
-    log: ((message: string) => void) & {
+    log: {
         error(message: string): void
         verbose(message: string): void
         warning(message: string): void
