@@ -1,13 +1,7 @@
 "use strict";
 
-import aws = require("aws-sdk");
 import shell = require("shelljs");
 import {TEMP_DIR} from "./base";
-
-// the us-west-2 is set in each test config
-const apiGateway = new aws.APIGateway({
-    region: "us-west-2"
-});
 
 /**
  * Executes given shell command.
@@ -38,71 +32,8 @@ async function createTempDir(tempDir, folderName) {
     await exec(`ln -s $(pwd) ${tempDir}/node_modules/`);
 
     await exec(`ln -s $(pwd)/node_modules/serverless ${tempDir}/node_modules/`);
-    // link serverless to the bin directory so we can use $(npm bin) to get the path to serverless
+    // we use npx running the local serverless in case not exists the global serverless will be used
     await exec(`ln -s $(pwd)/node_modules/serverless/bin/serverless.js ${tempDir}/node_modules/.bin/serverless`);
-}
-
-/**
- * Gets endpoint type of given URL from AWS
- * @param url
- * @returns {Promise<String>}
- */
-async function getEndpointType(url) {
-    const result = await apiGateway.getDomainName({
-        domainName: url,
-    }).promise();
-
-    return result.endpointConfiguration.types[0];
-}
-
-/**
- * Gets stage of given URL from AWS
- * @param url
- * @returns {Promise<String>}
- */
-async function getStage(url) {
-    const result = await apiGateway.getBasePathMappings({
-        domainName: url,
-    }).promise();
-
-    return result.items[0].stage;
-}
-
-/**
- * Gets basePath of given URL from AWS
- * @param url
- * @returns {Promise<String>}
- */
-async function getBasePath(url) {
-    const result = await apiGateway.getBasePathMappings({
-        domainName: url,
-    }).promise();
-
-    return result.items[0].basePath;
-}
-
-/**
- * Make API Gateway calls to create an API Gateway
- * @param {string} restApiName
- * @return {Object} Contains restApiId and resourceId
- */
-async function setupApiGatewayResources(restApiName) {
-    const restApiInfo = await apiGateway.createRestApi({name: restApiName}).promise();
-    const restApiId = restApiInfo.id;
-    const resourceInfo = await apiGateway.getResources({restApiId}).promise();
-    const resourceId = resourceInfo.items[0].id;
-    shell.env.REST_API_ID = restApiId;
-    shell.env.RESOURCE_ID = resourceId;
-    return {restApiId, resourceId};
-}
-
-/**
- * Make API Gateway calls to delete an API Gateway
- * @param {string} restApiId
- * @return {boolean} Returns true if deleted
- */
-async function deleteApiGatewayResources(restApiId) {
-    return apiGateway.deleteRestApi({restApiId}).promise();
 }
 
 /**
@@ -111,7 +42,7 @@ async function deleteApiGatewayResources(restApiId) {
  * @returns {Promise<void>}
  */
 function slsCreateDomain(tempDir) {
-    return exec(`cd ${tempDir} && $(npm bin)/serverless create_domain`);
+    return exec(`cd ${tempDir} && npx serverless create_domain`);
 }
 
 /**
@@ -121,7 +52,7 @@ function slsCreateDomain(tempDir) {
  * @returns {Promise<void>}
  */
 function slsDeploy(tempDir, debug: boolean = false) {
-    return exec(`cd ${tempDir} && $(npm bin)/serverless deploy` + (debug ? " --verbose" : ""));
+    return exec(`cd ${tempDir} && npx serverless deploy` + (debug ? " --verbose" : ""));
 }
 
 /**
@@ -130,7 +61,7 @@ function slsDeploy(tempDir, debug: boolean = false) {
  * @returns {Promise<void>}
  */
 function slsDeleteDomain(tempDir) {
-    return exec(`cd ${tempDir} && $(npm bin)/serverless delete_domain`);
+    return exec(`cd ${tempDir} && npx serverless delete_domain`);
 }
 
 /**
@@ -139,7 +70,7 @@ function slsDeleteDomain(tempDir) {
  * @returns {Promise<void>}
  */
 function slsRemove(tempDir) {
-    return exec(`cd ${tempDir} && $(npm bin)/serverless remove`);
+    return exec(`cd ${tempDir} && npx serverless remove`);
 }
 
 /**
@@ -188,9 +119,4 @@ export {
     slsDeploy,
     slsDeleteDomain,
     slsRemove,
-    getEndpointType,
-    getBasePath,
-    getStage,
-    setupApiGatewayResources,
-    deleteApiGatewayResources,
 };
